More implementation of validation to expectation iota proof checking
I am abandoning the attempt to prove the expected iota proofs are a subset, with the possibility of different iota naming, of the validation produced iota proofs. For starters, finding an allowable (Iota -> Iota) mapping is NP-hard, although possibly managable for the small inputs we would usually be dealing with.

But even worse, there may be equality or other relations that are expanded or collapsed in the expected proofs relative to the validation produced proofs. So the validation proofs may prove everthing we wish to check for, but, by virtue of being more or less verbose, there may be no possible Iota mapping. Similarly this may cause false apparent equivalences. Proving our expected iota proofs are a subset of the validation produced iota proofs is thus essentially generalized proof reduction which is NP-hard.

Instead we will only expect proofs in tests where all the terms are variables (either normal program variables or from proof assertions)